+++ lib/src/controllers/decorated_text_editing_controller.dart
@@ -27,6 +27,7 @@ class DecoratedTextEditingController extends TextEditingController {
       style: originalTextSpan.style!,
       source: text,
       rules: rules,
+      selectable: true,
     );
   }
 }