+++ lib/src/misc/functions.dart
@@ -9,6 +9,7 @@ TextSpan getDecoratedTextSpan({
   required String source,
   required TextStyle style,
   required List<DecoratorRule> rules,
+  bool selectable = false,
 }) {
   final decorations = Decorator(
     rules: rules,
@@ -32,7 +33,53 @@ TextSpan getDecoratedTextSpan({
               index,
               TextSpan(
                 style: item.rule?.style,
-                text: text,
+                text: selectable ||
+                        (item.rule?.leadingBuilder == null &&
+                            item.rule?.leadingBuilder == null)
+                    ? text
+                    : null,
+                children: selectable ||
+                        (item.rule?.leadingBuilder == null &&
+                            item.rule?.leadingBuilder == null)
+                    ? null
+                    : [
+                        WidgetSpan(
+                          alignment: PlaceholderAlignment.middle,
+                          child: GestureDetector(
+                            onTap: () {
+                              final decoration = decorations[index];
+                              if (decoration.rule?.onTap != null) {
+                                decoration.rule?.onTap!(
+                                  decoration.range.textInside(source).trim(),
+                                );
+                              }
+                            },
+                            child: Row(
+                              mainAxisSize: MainAxisSize.min,
+                              children: [
+                                if (item.rule?.leadingBuilder != null)
+                                  item.rule!.leadingBuilder!(
+                                    decorations[index]
+                                        .range
+                                        .textInside(source)
+                                        .trim(),
+                                  ),
+                                Text(
+                                  text,
+                                  style: item.rule?.style,
+                                ),
+                                if (item.rule?.trailingBuilder != null)
+                                  item.rule!.trailingBuilder!(
+                                    decorations[index]
+                                        .range
+                                        .textInside(source)
+                                        .trim(),
+                                  ),
+                              ],
+                            ),
+                          ),
+                        ),
+                      ],
                 recognizer: TapGestureRecognizer()
                   ..onTap = () {
                     final decoration = decorations[index];